Fix wrong field sizes in PGN 60928 ISO Address Claim.
The field sizes for both Device Instance fields are split 3+5 instead
of 4+4 bits.  Device Function is actually specified as 8 bits instead
of one byte.  There was an extraneous Reserved field and the trailing
Reserved field is now 1 instead of 3 bits, fixing the total length to
eight bytes.

Additionally, the Idustry Code field is actually called Industry Group
throughout the standard, so rename it accordingly.

Source of information for current PGN 60928 format is:
http://www.nmea.org/Assets/20140710%20nmea-2000-060928%20iso%20address%20claim%20pgn%20corrigendum.pdf