Fix message lengths for Airmar proprietary PGNs.
According to the quoted manuals, the PGN 126720 variants for Airmar
are all marked as "Single Frame: No".  Where the message lengths are
not greater than 8 bytes, add a comment to warn that the parser might
fail because it expects a single-frame format.

Use RES_NOTUSED for Reserved fields to suppress useless output.

Add a 126720-44 variant, as found in the DST200 User Manual document.
Extend the lookup for the PGN 126720-46 Transmission Interval field (3
means data not available).