Fix PGN 127493 Transmission Parameters, Dynamic.
Based on own research and the official PGN and field list at
http://www.nmea.org/Assets/july%202010%20nmea2000_v1-301_app_b_pgn_field_list.pdf

This PGN is transmitted with DLC=8 bytes and contains another Reserved
Bits field at the end. The Engine Instance is actually 8 bits wide.
The Transmission Gear field is padded with 6 bits to the next byte
boundary for the following 16-bit values.