Update PGN 129808 DSC call information
The prize for worst PGN of the month goes to the DSC PGN. It does not abstract away the underlying DSC implementation and use
nullable fixed size fields like real PGNs, but a combination of variable fields, strange lookup ranges (straight out of ITU M.821
and M.493) and overloaded fields. Some of the worst overloading is handled here by means of the mechanism used to decode
manufacturer proprietary PGNs.