Add Garmin and PCAN sample trace files.
These logs were recorded with overlapping timestamps on a Garmin
GPSmap 4008, a GPSmap 820 and simultaneously with a PCAN-USB dongle
from PEAK-Systems.

The GPSmap 820 was being connected to the bus which already had the
GPSmap 4008, a Garmin GBS10 battery sensor and a Garmin GHP12 / GHC20
auto pilot running.  So there should be lots of interesting Address
Claim, Product Information and standard NMEA PGNs in there.

Only thing that's missing is that analyzer can read the CSV files of
the Garmin chart plotters' diagnostic logs and the .trc file format
of the PCAN-View utility.