fix: PGNInfos from object to array
There are several PGNs that are in pgns.xml several
times as PGNInfos. The old pgns.json had the PGNInfos
converted to properties of a holder object, keyed by
the pgn number. This was valid JSON, but later properties
overwrote earlier ones and they were not available.

This commit changes the PGNInfo holder to be an array,
that allows access to the duplicates. Code that uses
pgns.json must be adjusted to account for this change.