Centralize and fix Off/On lookup definitions
Create a single macro that defines Off/On states. You could argue that
the existing No/Yes states should also be Off/On, but this would change
many older PGNs. This change only updates the values for recent PGN
changes so is not too onerous.

Also remove the lookup strings for the 2^2-1 and 2^2-2 values, e.g.
2 and 3: the two last values in a NMEA number field are always Error
and Unknown, we do not need to explicitly set these.

This also fixes an error in the old LOOKUP_YES_NO lookup: it erroneously
used a binary specification that was interpreted as a decimal value by
the code, so it used 10 decimal for Error and 11 decimal for Unknown,
so these would never be generated.