Issue #270: Auto-update of version in package.json
When updating the json versions make sure that the package.json
file contains the same version as the XML/JSON files.

A further improvement would be to automatically generate a version
number on the PGN definitions only, with a list of changes.
Given the fixed nature of the XML files this should not be impossible.