Fix #314 - INDIRECT_LOOKUP added for DEVICE_FUNCTION.
Change the entire spatial concept(*) for looking up enumerations.

Instead of using arrays that are then scanned, generate either large
switch statements (in analyzer) that we leave for the compiler to
optimize, or lists of callbacks (in analyzer-explain).

This is done using the magic of the C pre-processor combined with
repeated include of the same file multiple times: lookup.h is now
included five times, each for a different set of defines.