Issue #309: Add some Simrad commands, phase 1.
Started with proper analysis of AP48 commands.

The intention is to slowly add all 'operation' commands from the AP48
and all data coming NAC-3 and AC-42. This will allow development of
open source AP controllers.

'Commissioning' is considered out-of-scope. Although some users
seem to think this useful, I feel that it is dangerous to try and
emulate the control head / MFD to such an extent, and that this should
be left to the device manufacturer. Also, it is different for every
generation of AP computer so I also think the work/benefit ratio is too
low.