Issue #509: Fix Peukert exponent offset
The offset in the XML file was the straight value out of the C code, but
in the source code the offset was expressed in the raw value (applied
immediately after decoding the binary value) before the scale
(resolution) was applied. In the XML file this doesn't make sense.

As it happens, the Peukert exponent was the only field affected by this.